requirements.txt
# --- 1. Orchestration & State Management ---
langgraph           # สำหรับจัดการ Cyclic Graph และสถานะของ Agent 
langchain           # Framework หลักในการเชื่อมต่อส่วนประกอบต่างๆ 
langchain-core      # ส่วนประกอบพื้นฐานของ LangChain

# --- 2. LLMs (The Brain) ---
langchain-openai    # สำหรับเชื่อมต่อ GPT-4o และ DeepSeek (OpenAI Compatible) 
openai              # Client สำหรับเรียกใช้ API โดยตรง 

# --- 3. Sandbox & Execution ---
docker              # Docker SDK สำหรับสั่งรันรหัสในตู้คอนเทนเนอร์ 

# --- 4. Persistence & Knowledge Layer ---
chromadb            # Vector Database สำหรับเก็บและค้นหาประสบการณ์แก้บั๊ก 
psycopg2-binary     # สำหรับเชื่อมต่อ PostgreSQL เพื่อเก็บสถิติการใช้งาน 

# --- 5. Web API & Server ---
fastapi             # สำหรับสร้าง REST API ให้หน้าบ้านเรียกใช้
uvicorn             # ASGI Server สำหรับรัน FastAPI
pydantic            # สำหรับจัดการ Schema ของข้อมูล (DebugRequest, etc.)

# --- 6. Utilities ---
python-dotenv       # สำหรับโหลดค่า API Key จากไฟล์ .env 
requests            # สำหรับเรียกใช้ HTTP Request ภายใน Agent